6 лабораторная

+++ src/main/java/com/example/demo/Repository/City/CityRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface CityRepository {
     Iterable<City> findAll();
-    Optional<City> findById(int id);
+    Optional<City> findById(long id);
     City saveCity(City city);
     City updateCity(City city);
     void deleteCity(long id);
