+++ src/main/java/com/example/demo/Repository/Park/ParkRepository.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 public interface ParkRepository {
     Iterable<Park> findAll();
-    Optional<Park> findById(int id);
+    Optional<Park> findById(long id);
     Park savePark(Park park);
     Park updatePark(Park park);
     void deletePark(long id);
